send a mail when the engine crashes - closes #207

+++ src/core/path.cpp
@@ -21,7 +21,7 @@ namespace Lumix
 	{
 		size_t len = strlen(path);
 		ASSERT(len < LUMIX_MAX_PATH);
-		PathUtils::normalize(path, m_path, len + 1);
+		PathUtils::normalize(path, m_path, (uint32_t)len + 1);
 		m_id = crc32(m_path);
 	}
